Identified Issue with Vuforia Test
When using the camera stream on Vuforia test, the image does not show the x, y, and z axes on the skystone image. However, the camera still identifies the image as our if statement to go forward when the image is detected still runs. We copied the vuforia template provided by FTC and pasted it into the VuforiaTesting class and the axes were displayed. Hence, for tomorrow we will be cross referencing the two to see where they are different.